+++ ext/jsonrpclib/config.py
@@ -3,13 +3,13 @@
 """
 The configuration module.
 
-:copyright: Copyright 2019, Thomas Calmant
+:copyright: Copyright 2020, Thomas Calmant
 :license: Apache License 2.0
-:version: 0.4.0
+:version: 0.4.1
 
 ..
 
-    Copyright 2019 Thomas Calmant
+    Copyright 2020 Thomas Calmant
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -29,7 +29,7 @@ import sys
 # ------------------------------------------------------------------------------
 
 # Module version
-__version_info__ = (0, 4, 0)
+__version_info__ = (0, 4, 1)
 __version__ = ".".join(str(x) for x in __version_info__)
 
 # Documentation strings format