+++ themes-default/slim/src/components/show-header.vue
@@ -167,9 +167,9 @@
                                     <tr><td class="showLegend">Default EP Status: </td><td>{{ show.config.defaultEpisodeStatus }}</td></tr>
                                     <tr><td class="showLegend"><span :class="{'invalid-value': !show.config.locationValid}">Location: </span></td><td><span :class="{'invalid-value': !show.config.locationValid}">{{show.config.location}}</span>{{show.config.locationValid ? '' : ' (Missing)'}}</td></tr>
 
-                                    <tr v-if="show.config.aliases.length > 0">
+                                    <tr v-if="show.config.aliases.filter(alias => alias.season === -1).length > 0">
                                         <td class="showLegend" style="vertical-align: top;">Scene Name:</td>
-                                        <td>{{show.config.aliases.join(', ')}}</td>
+                                        <td>{{show.config.aliases.filter(alias => alias.season === -1).map(alias => alias.title).join(', ')}}</td>
                                     </tr>
 
                                     <tr v-if="show.config.release.requiredWords.length + search.filters.required.length > 0">
@@ -501,6 +501,7 @@ export default {
         },
         changeStatusOptions() {
             const { search, getStatus, statuses } = this;
+            const { general } = search;
 
             if (statuses.length === 0) {
                 return [];
@@ -510,7 +511,7 @@ export default {
             const defaultOptions = ['wanted', 'skipped', 'ignored', 'downloaded', 'archived']
                 .map(key => getStatus({ key }));
 
-            if (search.useFailedDownloads) {
+            if (general.failedDownloads.enabled) {
                 defaultOptions.push(getStatus({ key: 'failed' }));
             }