+++ themes-default/slim/src/store/mutation-types.js
@@ -16,6 +16,8 @@ const ADD_CONFIG = '⚙️ Config added to store';
 const ADD_SHOW = '📺 Show added to store';
 const ADD_SHOW_EPISODE = '📺 Shows season with episodes added to store';
 const ADD_STATS = 'ℹ️ Statistics added to store';
+const ADD_SHOW_SCENE_EXCEPTION = '📺 Add a scene exception';
+const REMOVE_SHOW_SCENE_EXCEPTION = '📺 Remove a scene exception';
 
 export {
     LOGIN_PENDING,
@@ -35,5 +37,7 @@ export {
     ADD_CONFIG,
     ADD_SHOW,
     ADD_SHOW_EPISODE,
-    ADD_STATS
+    ADD_STATS,
+    ADD_SHOW_SCENE_EXCEPTION,
+    REMOVE_SHOW_SCENE_EXCEPTION
 };